restructured test directory, fixed some import statements elsewhere
