+++ medacy/pipelines/systematic_review_pipeline.py
@@ -4,8 +4,8 @@ import spacy
 from medacy.pipelines.base.base_pipeline import BasePipeline
 from medacy.pipeline_components.feature_overlayers.gold_annotator_component import GoldAnnotatorComponent
 from medacy.pipeline_components.feature_extracters.discrete_feature_extractor import FeatureExtractor
-from medacy.pipeline_components.feature_overlayers.metamap import MetaMap
-from medacy.pipeline_components.feature_overlayers.metamap import MetaMapComponent
+from medacy.pipeline_components.feature_overlayers.metamap.metamap import MetaMap
+from medacy.pipeline_components.feature_overlayers.metamap.metamap_component import MetaMapComponent
 from medacy.pipeline_components.tokenizers.systematic_review_tokenizer import SystematicReviewTokenizer
