+++ medacy/tests/pipeline_components/learners/test_bilstm_crf.py
@@ -8,7 +8,7 @@ import pkg_resources
 
 from medacy.data.dataset import Dataset
 from medacy.model import Model
-from medacy.pipelines import LstmSystematicReviewPipeline
+from medacy.pipelines.lstm_systematic_review_pipeline import LstmSystematicReviewPipeline
 from medacy.data.annotations import Annotations
 
 
@@ -25,20 +25,17 @@ class TestBiLstmCrf(TestCase):
 
         cls.train_dataset, _,  meta_data = Dataset.load_external('medacy_dataset_end')
         cls.entities = meta_data['entities']
-        cls.train_dataset.set_data_limit(1)
+        cls.train_dataset.data_limit = 1
 
         cls.test_dataset, _, _ = Dataset.load_external('medacy_dataset_end')
-        cls.test_dataset.set_data_limit(2)
+        cls.test_dataset.data_limit = 2
 
-        cls.prediction_directory = tempfile.mkdtemp() #directory to store predictions
-
-
+        cls.prediction_directory = tempfile.mkdtemp()  # directory to store predictions
 
     @classmethod
     def tearDownClass(cls):
         pkg_resources.cleanup_resources()
         shutil.rmtree(cls.prediction_directory)
-
 
     def test_prediction_with_testing_pipeline(self):
         """
@@ -48,15 +45,15 @@ class TestBiLstmCrf(TestCase):
 
         pipeline = LstmSystematicReviewPipeline(
             entities=['tradename'],
-            word_embeddings='medacy/tests/ner/model/test_word_embeddings.txt',
+            word_embeddings='medacy/tests/sample_data/test_word_embeddings.txt',
             cuda_device=-1
         )
 
-        #train on Abelcet.ann
+        # train on Abelcet.ann
         model = Model(pipeline)
         model.fit(self.train_dataset)
 
-        #predict on both
+        # predict on both
         model.predict(self.test_dataset, prediction_directory=self.prediction_directory)
 
         second_ann_file = "%s.ann" % self.test_dataset.get_data_files()[1].file_name
